+++ examples/builder/internal-seekable.rb
@@ -10,16 +10,14 @@ JSONParser =
   .matcher do |input, **_options|
     File.extname(input) == '.json'
   end
-  .caller do |input, **options, &block|
+  .caller do |input, **options, &original|
     File.open(input) do |file|
       json = ::JSON.load(file)
       # The following block call is equivalent to calling `super` in the class style.
-      block.call(json, **options)
+      original.call(json, **options)
     end
   end
-  .terminator do
-    true
-  end
+  .terminator { true }
   .build
 
 XMLParser =
@@ -28,27 +26,24 @@ XMLParser =
   .matcher do |input, **_options|
     File.extname(input) == '.xml'
   end
-  .caller do |input, **options, &block|
+  .caller do |input, **options, &original|
     File.open(input) do |file|
       # The following block call is equivalent to calling `super` in the class style.
-      block.call(REXML::Document.new(file), **options)
+      original.call(REXML::Document.new(file), **options)
     end
   end
-  .terminator do
-    true
-  end
+  .terminator { true }
   .build
 
 def build_json_scraper(type)
   CallableTree::Node::External::Builder
     .new
     .matcher do |input, **_options|
-      !!input[type.to_s]
+      !input[type.to_s].nil?
     end
     .caller do |input, **_options|
       input[type.to_s]
-        .map { |element| [element['name'], element['emoji']] }
-        .to_h
+        .to_h { |element| [element['name'], element['emoji']] }
     end
     .build
 end
@@ -66,8 +61,7 @@ def build_xml_scraper(type)
       input
         .get_elements("//#{type}")
         .first
-        .map { |element| [element['name'], element['emoji']] }
-        .to_h
+        .to_h { |element| [element['name'], element['emoji']] }
     end
     .build
 end