+++ lib/callable_tree/node/internal/strategy/broadcast.rb
@@ -7,9 +7,26 @@ module CallableTree
         class Broadcast
           include Strategy
 
+          def initialize(terminable: false)
+            self.terminable = terminable
+            @terminator =
+              if terminable
+                proc { |node, output, *inputs, **options| node.terminate?(output, *inputs, **options) }
+              else
+                proc { false }
+              end
+          end
+
           def call(nodes, *inputs, **options)
-            nodes.map do |node|
-              node.call(*inputs, **options) if node.match?(*inputs, **options)
+            nodes.reduce([]) do |outputs, node|
+              output = (node.call(*inputs, **options) if node.match?(*inputs, **options))
+              outputs << output
+
+              if @terminator.call(node, output, *inputs, **options)
+                break outputs
+              else
+                outputs
+              end
             end
           end
         end