+++ lib/callable_tree/node/internal/strategy/seek.rb
@@ -7,13 +7,23 @@ module CallableTree
         class Seek
           include Strategy
 
+          def initialize(terminable: true)
+            self.terminable = terminable
+            @terminator =
+              if terminable
+                proc { |node, output, *inputs, **options| node.terminate?(output, *inputs, **options) }
+              else
+                proc { false }
+              end
+          end
+
           def call(nodes, *inputs, **options)
             nodes
               .lazy
               .select { |node| node.match?(*inputs, **options) }
               .map do |node|
                 output = node.call(*inputs, **options)
-                terminated = node.terminate?(output, *inputs, **options)
+                terminated = @terminator.call(node, output, *inputs, **options)
                 [output, terminated]
               end
               .select { |_output, terminated| terminated }