+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## [Unreleased]
 
+## [0.3.9] - 2022-11-06
+
+- Change `CallableTree::Node::Internal#broadcastable` to take `terminable` keyword parameter as boolean. It defaults to `false`, which is the same behavior as before.
+- Change `CallableTree::Node::Internal#composable` to take `terminable` keyword parameter as boolean. It defaults to `false`, which is the same behavior as before.
+- Change `CallableTree::Node::Internal#seekable` to take `terminable` keyword parameter as boolean. It defaults to `true`, which is the same behavior as before.
+
 ## [0.3.8] - 2022-05-05
 
 - (Experimental) Add `CallableTree::Node::Internal::Builder#identifier`.